+++ toolchain/check/sem_ir_loc_diagnostic_emitter.h
